teacher

+++ my-app/src/main/java/at/ac/htl/bhitm3/model/Teacher.java
@@ -1,8 +1,9 @@
 package at.ac.htl.bhitm3.model;
 
 public class Teacher extends Person {
-    double salary;
-    public void setSalary(double salary) {
+    int salary;
+   
+    public void setSalary(int salary) {
         this.salary = salary;
     }
